LPS-122549 Hide usages of other display contexts
